Split office providers into one class per file
